+++ computenest-20210601/alibabacloud_computenest20210601/models/_generate_service_policy_request.py
@@ -17,7 +17,7 @@ class GenerateServicePolicyRequest(DaraModel):
         template_name: str = None,
         trial_type: str = None,
     ):
-        # The types of operations for which to generate policy information.
+        # The operation types for which policy information needs to be generated.
         self.operation_types = operation_types
         self.parameters = parameters
         # The region ID.
@@ -32,11 +32,11 @@ class GenerateServicePolicyRequest(DaraModel):
         self.service_version = service_version
         # The template name.
         self.template_name = template_name
-        # The trial type. The default value is NotTrial. Valid values:
+        # The trial type. Default value: NotTrial. Valid values:
         # 
-        # - Trial: The service supports trial use.
+        # - Trial: Trial is supported.
         # 
-        # - NotTrial: The service does not support trial use.
+        # - NotTrial: Trial is not supported.
         self.trial_type = trial_type
 
     def validate(self):