+++ computenest-20210601/alibabacloud_computenest20210601/models/_get_skill_request.py
@@ -9,7 +9,7 @@ class GetSkillRequest(DaraModel):
         self,
         skill_id: str = None,
     ):
-        # The ID of the Skill.
+        # The unique ID of the Skill.
         # 
         # This parameter is required.
         self.skill_id = skill_id