+++ computenest-20210601/alibabacloud_computenest20210601/models/_list_skills_request.py
@@ -15,13 +15,13 @@ class ListSkillsRequest(DaraModel):
         need_download_url: bool = None,
         next_token: str = None,
     ):
-        # The filters for querying Skills.
+        # The filter.
         self.filter = filter
-        # The maximum number of entries to return per page.
+        # The maximum number of entries to return in this request.
         self.max_results = max_results
-        # Specifies whether to return the download URL of the Skill package.
+        # Specifies whether to include the download URL of the skill package.
         self.need_download_url = need_download_url
-        # The token for the next page of results. Leave this parameter empty for the first request. For subsequent requests, use the `NextToken` value from the previous response.
+        # NextToken
         self.next_token = next_token
 
     def validate(self):
@@ -76,23 +76,23 @@ class ListSkillsRequestFilter(DaraModel):
         name: str = None,
         value: List[str] = None,
     ):
-        # The filter name. Valid values:
+        # The filter name. You can specify one or more names for the query. Valid values:
         # 
-        # - `SkillId`: The Skill ID. An exact match is performed.
+        # - SkillId: the skill ID. Exact match.
         # 
-        # - `SkillSpaceId`: The ID of the SkillSpace. An exact match is performed.
+        # - SkillSpaceId: the SkillSpace ID. Exact match.
         # 
-        # - `SkillName`: The Skill name.
+        # - SkillName: the skill name.
         # 
-        # - `MatchType`: The match type for `SkillName`. Valid values: `exact` (exact match), `prefix` (prefix match), and `fuzzy` (fuzzy match).
+        # - MatchType: the match type that controls how SkillName is matched. Valid values: exact, prefix, and fuzzy.
         # 
-        # - `SkillType`: The Skill type. Valid values: `official` and `custom`.
+        # - SkillType: the skill type. Valid values: official and custom.
         # 
-        # - `Keyword`: The keyword for a fuzzy match on the Skill name or Skill description.
+        # - Keyword: keyword match for the skill name or skill description. Fuzzy match.
         # 
-        # - `SkillLabels`: The Skill labels. A fuzzy match is performed.
+        # - SkillLabels: the skill labels. Fuzzy match.
         self.name = name
-        # The filter values. You can specify a maximum of 10 values.
+        # The list of filter values. Valid values of N: 1 to 10.
         self.value = value
 
     def validate(self):