+++ computenest-20210601/ChangeLog.md
@@ -1,3 +1,14 @@
+2026-07-02 Version: 4.5.2
+- Update API CreateSkill: add request parameters SkillDisplayName.
+- Update API GetSkill: add response parameters Body.Locales.
+- Update API GetSkill: add response parameters Body.SkillDisplayName.
+- Update API ListPublicSkills: add response parameters Body.Skills.$.Locales.
+- Update API ListPublicSkills: add response parameters Body.Skills.$.SkillDisplayName.
+- Update API ListSkills: add response parameters Body.Skills.$.Locales.
+- Update API ListSkills: add response parameters Body.Skills.$.SkillDisplayName.
+- Update API UpdateSkill: add request parameters SkillDisplayName.
+
+
 2026-06-12 Version: 4.5.1
 - Update API GenerateServicePolicy: add request parameters Parameters.